refactor(firebase): abstracts get user function
